requirements.txt
tortoise-orm
annoy
crewai
aerich
psycopg2
asyncpg
tortoise-orm[psycopg]
fastapi
uvicorn
python-decouple
python-dotenv
crewai[tools]
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
redis 
aioredis